+++ app/src/managers/plugins/range/RangeManager.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import type { CoreState } from '../../../timepicker/CoreState';
 import type { EventEmitter, TimepickerEventMap } from '../../../utils/EventEmitter';
 import type { RangeMinuteCommitEventData, ConfirmEventData } from '../../../types/types';