+++ app/webpack.config.js
@@ -2,6 +2,7 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+import ESLintPlugin from 'eslint-webpack-plugin';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -108,6 +109,12 @@ export default {
     }),
     new MiniCssExtractPlugin({
       filename: '[name].css',
+    }),
+    new ESLintPlugin({
+      extensions: ['ts'],
+      files: 'src/**/*.ts',
+      emitWarning: true,
+      failOnError: false,
     }),
   ],
   performance: {